+++ samples/Digipolis.Web.SampleApi/Logic/Mapping/AutoMapperProfileConfiguration.cs
@@ -12,8 +12,6 @@ namespace Digipolis.Web.SampleApi.Logic.Mapping
         {
             CreateMap<Value, ValueDto>().ReverseMap();
             CreateMap<ValueType, ValueTypeDto>().ReverseMap();
-            CreateMap(typeof(PagedResult<>), typeof(PagedResult<>));
-            CreateMap(typeof(IEnumerable<>), typeof(PagedResult<>)).ForMember("Data", x => x.MapFrom(m => m));
         }
                
     }