+++ samples/Digipolis.Web.SampleApi/Models/CriteriaDto.cs
@@ -1,6 +1,7 @@
 ﻿using Digipolis.Web.Api;
 using Digipolis.Web.Modelbinders;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Infrastructure;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -10,7 +11,6 @@ namespace Digipolis.Web.SampleApi.Models
 {
     public class CriteriaDto : PageSortOptions
     {
-
         public bool? TestBool { get; set; }
        
         public string[] StringArray { get; set; }