+++ samples/Digipolis.Web.SampleApi/Models/EmbeddedValueDto.cs
@@ -1,0 +1,12 @@
+﻿using Digipolis.Web.Api;
+using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace Digipolis.Web.SampleApi.Models
+{
+    public class EmbeddedValueDto : Embedded<ValueDto>
+    {
+        [JsonProperty("valueDtos")]
+        public override IEnumerable<ValueDto> ResourceList { get; set; }
+    }
+}