+++ src/Digipolis.Web/Swagger/SwaggerSettings.cs
@@ -21,8 +21,8 @@ namespace Digipolis.Web.Swagger
 
             options.OperationFilter<AddFileUploadParams>();
             options.OperationFilter<AddConsumeProducesValues>();
+            options.OperationFilter<ValidRefUri>();
             options.DocumentFilter<SetVersionInPaths>();
-            options.SchemaFilter<PagedResultSchemaFilter>();
             options.DocumentFilter<EndPointPathsAndParamsToLower>();
 
             Configuration(options);