+++ test/Digipolis.Web.UnitTests/Api/JsonConverters/PagedResultConverterTests.cs
@@ -13,13 +13,12 @@ namespace Digipolis.Web.UnitTests.Api.JsonConverters
         public void CanBeDeserialized()
         {
             var settings = new JsonSerializerSettings();
-            settings.Converters.Add(new PagedResultConverter());
 
             var original = new PagedResult<DummyType>(1, 10, 25, new List<DummyType> { new DummyType { Text = "someText" } });
 
             var serialized = JsonConvert.SerializeObject(original, settings);
 
-            var deserialized = (PagedResult<DummyType>)JsonConvert.DeserializeObject<DeserializationPagedResult<DummyType>>(serialized, settings);
+            var deserialized = JsonConvert.DeserializeObject<PagedResult<DummyType>>(serialized, settings);
 
             var deserialized2 = JsonConvert.DeserializeObject<PagedResult<DummyType>>(serialized, settings);
 
@@ -28,8 +27,8 @@ namespace Digipolis.Web.UnitTests.Api.JsonConverters
             Assert.Equal(1, deserialized.Page.Size);
             Assert.Equal(25, deserialized.Page.TotalElements);
             Assert.Equal(3, deserialized.Page.TotalPages);
-            Assert.NotNull(deserialized.Data);
-            Assert.Equal("someText", deserialized.Data.First().Text);
+            Assert.NotNull(deserialized.Embedded?.ResourceList);
+            Assert.Equal("someText", deserialized.Embedded.ResourceList.First().Text);
         }
 
         public class DummyType